+++ STS2AIAgent.Tests/TestRunner.cs
@@ -522,6 +522,9 @@ internal static class TestRunner
         yield return ("SourceShape.BudgetsTrackTheirFiles", () => Task.Run(SourceShapeContractTests.BudgetsStayCloseToTheFilesTheyGuard));
         yield return ("ActionDiagnostics.NoWordlessRecoveryCatch", () => Task.Run(ActionDiagnosticsContractTests.NoRecoveryCatchSwallowsWithoutSayingSo));
         yield return ("ActionDiagnostics.FaultedTasksNameTheirException", () => Task.Run(ActionDiagnosticsContractTests.FaultedGameTasksNameTheirException));
+        yield return ("MonsterMoves.FromTheLocTableDirectly", () => Task.Run(MonsterMovesExportContractTests.MovesComeFromTheLocalizationTableDirectly));
+        yield return ("MonsterMoves.OnlyTitles", () => Task.Run(MonsterMovesExportContractTests.OnlyTheTitleOfEachMoveIsExported));
+        yield return ("MonsterMoves.PrefixFromTitleKey", () => Task.Run(MonsterMovesExportContractTests.ThePrefixFollowsTheMonstersOwnTitleKey));
         yield return ("AbandonRun.StopsAtTheConfirmation", () => Task.Run(AbandonRunContractTests.AbandonRunStopsAtTheConfirmation));
         yield return ("AbandonRun.RefusesRatherThanClickingBlind", () => Task.Run(AbandonRunContractTests.AbandonRunRefusesRatherThanClickingBlind));
         yield return ("ReflectedMembers.NoLookupOutsideTheRegistry", () => Task.Run(ReflectedMemberRegistryTests.NothingLooksUpAGameMemberByNameOutsideTheRegistry));